requirements.txt
# 核心依赖
langchain-core>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-experimental>=0.0.55
langgraph>=0.0.26

# 向量存储和嵌入
langchain-chroma>=0.0.12
langchain-ollama>=0.1.0
chromadb>=0.4.24

# LLM集成
langchain-deepseek>=0.0.3
openai>=1.0.0

# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# 数据模型和验证
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 环境配置
python-dotenv>=1.0.0

# 工具和工具类
langchain-text-splitters>=0.0.1

# 文档处理
unstructured>=0.12.0
pdfminer-six>=20221105
docx2txt>=0.8

# 异步和工具
asyncio>=3.4.3
aiofiles>=23.2.0

# 开发依赖（可选）
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0